Improve Huffman encoding/decoding process.

+++ Compressor/Compressors/LZW.h
@@ -24,7 +24,7 @@ public:
 	/**
 	 *
 	 */
-	void encode(const vector<uchar>& data, vector<uchar>& outputData);
+	void encode(const vector<uchar>& data, vector<int>& outputData);
 
 private:
 	/**
